+++ bugzoo/__init__.py
@@ -7,3 +7,4 @@ from bugzoo.core.bug import Bug
 
 import logging
 logging.getLogger(__name__).addHandler(logging.NullHandler())
+logging.getLogger(__name__).setLevel(logging.INFO)