+++ bugzoo/cli/__init__.py
@@ -1,5 +1,6 @@
 from typing import List, Optional, Dict
 import sys
+import os
 import argparse
 import logging
 from operator import itemgetter
@@ -528,10 +529,11 @@ def build_parser():
 def main():
     try:
         log_formatter = logging.Formatter('%(levelname)s:%(name)s:%(asctime)s: %(message)s')  # noqa: pycodestyle
-        log_to_stdout = logging.StreamHandler()
-        log_to_stdout.setFormatter(log_formatter)
-        logging.getLogger('bugzoo').setLevel(logging.INFO)
-        logging.getLogger('bugzoo').addHandler(log_to_stdout)
+        log_filename = os.path.join(os.getcwd(), "bugzood.log")
+        log_to_file = logging.handlers.WatchedFileHandler(log_filename)
+        log_to_file.setFormatter(log_formatter)
+        logging.getLogger('bugzoo').setLevel(logging.DEBUG)
+        logging.getLogger('bugzoo').addHandler(log_to_file)
 
         parser = build_parser()
         args = parser.parse_args()